Experimental Inline Malloy editor (#303)
A substantial change that allows a stable query struct in the visual editor to be converted to a text based Malloy document and edited in situ. If the changes made with in text mode are stable query compatible, they can be converted back into a stable query.

The main caveats are:

* Bring your own Monaco editor. Because of the complexities of bringing up a Monaco Javascript Worker (or not) in various environments, that's left as an exercise to the embedder.
* For now the Malloy language server portion of the editor requires a Malloy ModelDef object, instead of a ModelInfo object like the rest of the editor. There are convenience routines to create a ModelInfo from a ModelDef so you should only have to compile one.